+++ python/exp04_update_of_c.py
@@ -62,9 +62,9 @@ def syr2k_implicit_compact(A, B, C):
 
 def exp04_update_of_c(b, n):
 
-    A = torch.randn(n, n)
-    B = torch.randn(n, n)
-    C = torch.randn(n, n)
+    A = torch.randn((n, n), dtype=torch.float64)
+    B = torch.randn((n, n), dtype=torch.float64)
+    C = torch.randn((n, n), dtype=torch.float64)
 
     b.benchmark('add', add, A, B)
     b.benchmark('scal', scal, A)
